refactor: rename baker-openapi-transportation -> baker-openapi-wirespec
The module is the wirespec-integration runtime for baker-openapi (it
bridges Wirespec.RawRequest/RawResponse with a default HTTP client and
will host the default handler-building behaviour next), so naming it
after wirespec rather than 'transportation' reflects what it owns.
